feat(EIDOMNI-210): refactor DcqlUtil and related tests to new package structure
